+++ CHANGELOG.md
@@ -19,6 +19,7 @@ BUG FIXES:
 
 - **netapp-ontap_cluster_peer**: fixed issue where cluster peering between GCNV and ONTAP failed because both sides were using the same `ipspace`. Added `peer_ipspace` attribute to allow specifying separate IPspaces for local and peer clusters. ([#682](https://github.com/NetApp/terraform-provider-netapp-ontap/issues/682))([#363](https://github.com/NetApp/terraform-provider-netapp-ontap/issues/363))
 - **netapp-ontap_volume resource**: Rectified validator for `analytics.state`. ([#734](https://github.com/NetApp/terraform-provider-netapp-ontap/issues/734))
+- **netapp-ontap_cluster_peer resource**: Fixed issue with import of source details by setting the option as optional. ([#744](https://github.com/NetApp/terraform-provider-netapp-ontap/issues/744))
 
 # 2.7.1 (2026-07-30)